Intermediate step
Whilst this is working, I have detected that I'm on the wrong way. Going
ahead with the following premisses:
The cluster shall get a public function synchronize() that stops all
measurement cycles of instruments in the cluster, sets instrument times
to UTC, stores the reference time as property of the SaradCluster object
and re-starts the measuring cycles.  The first set of measuring values
can only be read after the first cycle is finished.