!: String Representation not for instruments not working
It broke due to a change of "serial" changed to list instead of a dictionary.

Removing it from output since a list of "possible working baud rates" is not useful at first glance.

Also: fixed a global regex modifier "(?i)" being not in front of the regex. Changed from "deprecated" to "not supported" in a recent python version